Dev: unittest: move watchdog unit test code into test_watchdog.py
